fix: move all new helpers to utilities
