+++ src/utilities/treePluginEnhacer.js
